requirements.txt
pandas
numpy
matplotlib
xgboost
plotly
pillow
streamlit
scikit-learn
